requirements.txt
# SID-STACK: Python 3.11+, Chainlit UI, OpenAI-compatible client for OpenRouter
chainlit==2.11.0
openai==2.31.0
pyyaml==6.0.2
python-dotenv==1.0.1
gspread==6.1.4
